ineedyou/full_stack_web/backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<5.0
python-decouple>=3.8,<4.0
gunicorn>=21.2,<22.0
python-dotenv>=1.0,<2.0
drf-yasg>=1.21.7  # Swagger/OpenAPI documentation